requirements.txt
astroid==2.9.3
backports.entry-points-selectable==1.1.1
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
distlib==0.3.4
filelock>=3.4.0
firebase-admin==5.2.0
gcloud==0.18.3
google-api-core==1.31.3
google-api-python-client==1.12.3
google-auth==1.35.0
google-auth-httplib2==0.0.4
google-cloud-appengine-logging==0.2.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-logging==2.6.0
google-cloud-storage==1.43.0
google-crc32c==1.1.1
google-resumable-media==2.1.0
googleapis-common-protos==1.52.0
grpcio==1.43.0
grpcio-tools==1.43.0
gunicorn==20.1.0
httplib2==0.20.2
idna==2.10
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
msgpack==1.0.3
oauth2client==4.1.3
packaging==21.3
platformdirs>=2.4.0
proto-plus==1.19.8
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pylint==2.12.2
pyparsing==3.0.6
python-dotenv==0.19.1
pytz==2021.3
requests==2.23.0
rsa==4.8
six==1.16.0
toml==0.10.2
twilio==7.5.0
typing_extensions==4.0.1
uritemplate==3.0.1
urllib3==1.25.11
virtualenv==20.10.0
wrapt==1.13.3
Werkzeug==2.0.2
Jinja2==3.0.3
MarkupSafe==2.0.1
click==8.0.3
flask==2.0.2 
itsdangerous==2.0.1